feat(relayer): create fee interfaces file
