+++ packages/relayer/src/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from "./env";
 export * from "./ref";
+export * from './fee';
 export * from "./relayer";
